requirements.txt
streamlit>=1.24.0
plotly>=5.13.0
pandas>=1.5.3
numpy>=1.24.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
requests>=2.28.2
python-dotenv>=1.0.0
typing-extensions>=4.5.0

